% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{rs_stickr_list}
\alias{rs_stickr_list}
\title{List available stickers}
\usage{
rs_stickr_list(pic_type = "png")
}
\arguments{
\item{pic_type}{one of "png", "svg"}
}
\value{
character vector of sticker names
}
\description{
List available stickers
}
\examples{
rs_stickr_list()
}
