% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{stickr_insert}
\alias{stickr_insert}
\title{Add hex sticker in Rmd doc}
\usage{
stickr_insert(name, ...)
}
\arguments{
\item{name}{name of sticker, character of length 1}

\item{...}{other arguments passed to \code{\link{stickr_get}} and \code{\link[knitr:include_graphics]{?knitr::include_graphics}}, consider \code{dpi}}
}
\value{
call to \code{\link[knitr:include_graphics]{?knitr::include_graphics}}
}
\description{
Using the \code{dpi} argument in \code{...} will help scale width. Higher
dpi will correspond to smaller images in rendered documents. In some quick
testing, a dpi of 5000 results in an approximately in-line sized image.
}
\details{
If you hit rate limits on github API calls, consider setting the GITHUB_PAT environment variable.
See \code{\link[gh:gh]{?gh::gh}} for details.
}
\examples{
# This returns a function, really only makes sense inside Rmd
stickr_insert("R6")
}
