% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{stickr_list}
\alias{stickr_list}
\title{List available stickers}
\usage{
stickr_list(pic_type = "png")
}
\arguments{
\item{pic_type}{one of "svg" or "png", defaults to "png"}
}
\value{
character vector of sticker names
}
\description{
If you hit rate limits on github API calls, consider setting the GITHUB_PAT environment variable.
See \code{\link[gh:gh]{?gh::gh}} for details.
}
\examples{
stickr_list()
}
