% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{stickr_view}
\alias{stickr_view}
\title{View RStudio Stickers}
\usage{
stickr_view(name, ...)
}
\arguments{
\item{name}{name of sticker, character of length 1}

\item{...}{arguments passed to \code{\link{stickr_get}}}
}
\value{
file where sticker is, invisibly
}
\description{
View RStudio Stickers
}
\examples{
stickr_view("R6")
}
