% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{stickr_get}
\alias{stickr_get}
\title{Get hex sticker from web}
\usage{
stickr_get(
  name,
  destfile = NULL,
  view = TRUE,
  filename = NULL,
  path = NULL,
  repo = NULL
)
}
\arguments{
\item{name}{name of sticker, character of length 1}

\item{destfile}{destination, defaults to \code{NULL}}

\item{view}{show sticker after loading? Defaults to \code{TRUE}}

\item{filename}{filename of sticker in repository, use if auto-detection fails}

\item{path}{path within repository, use if auto-detection fails}

\item{repo}{repository name, use if auto-detection fails}
}
\value{
path to downloaded image
}
\description{
This function \emph{tries} to search for the sticker for the package you want.
}
\details{
It will automatically get stickers for any packages in the
\href{https://github.com/rstudio/hex-stickers}{rstudio/hex-stickers} repository
or if the package's GitHub repository is specified in the DESCRIPTION and
the \code{man/figures} directory contains (1) exactly one file
or (2) a file named \code{logo.png}. These cases cover \emph{most} CRAN packages.

For packages where the sticker cannot be automatically found or defaults must
be overridden, the \code{repository}, \code{path}, and \code{filename} arguments provide
escape hatches.

Packages stored somewhere other than GitHub are currently unsupported.

If you hit rate limits on GitHub API calls, consider setting the \code{GITHUB_PAT} environment variable.
See \code{\link[gh:gh]{gh::gh()}} for details.
}
\examples{
# Get a sticker
if(interactive()) r6 <- stickr_get("R6")
# Get a sticker in svg
if(interactive()) r6 <- stickr_get("R6", filename = "R6.svg")
# Get a particular sticker in the man/figures folder
if(interactive()) tm <- stickr_get("textmineR", filename = "textmineR_v8.png")
}
