% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{stickr_get}
\alias{stickr_get}
\title{Get hex sticker from web}
\usage{
stickr_get(name, pic_type = "png", destfile = NULL, view = TRUE)
}
\arguments{
\item{name}{name of sticker, character of length 1}

\item{pic_type}{one of "svg" or "png", defaults to "png"}

\item{destfile}{destination, defaults to NULL}

\item{view}{show sticker after loading? Defaults to TRUE}
}
\value{
path to downloaded image
}
\description{
If you hit rate limits on github API calls, consider setting the GITHUB_PAT environment variable.
See \code{\link[gh:gh]{?gh::gh}} for details.
}
\examples{
r6 <- stickr_get("R6")
}
