% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{stickr_view}
\alias{stickr_view}
\title{View RStudio Stickers}
\usage{
stickr_view(name, pic_type = "png")
}
\arguments{
\item{name}{name of sticker, character of length 1}

\item{pic_type}{one of "svg" or "png", defaults to "png"}
}
\value{
file where sticker is, invisibly
}
\description{
If you hit rate limits on github API calls, consider setting the GITHUB_PAT environment variable.
See \code{\link[gh:gh]{?gh::gh}} for details.
}
\examples{
stickr_view("R6")
}
